fix: update fetching resume logic
